Implement challenge0
Challenge 0 is a simple echo server: input in, same output out.

For upcoming challenges I found myself implementing the same pattern for
TCP servers: 1 process that listens and accepts connections, and 1
process per connection. It's a bit over the top for an echo server but
it makes the final implementation very simple.

The echo server is a `Task` which isn't great for long-running processes
(e.g. message handling is not expected) but this challenge is far from
long-running.