Implement challenge1
Challenge 1 is a server that can tell you if a number is prime.

Unlike challenge 0 the protocol is line-based: each request is expected
to end in a newline. Fortunately `:gen_tcp` and `:inet` support what a
"packet" mode of `:line` which results in a `:gen_tcp.recv\2` returning
only if a newline is detected. There is one catch to this: the receive
buffer must be large enough to hold the packet. Otherwise you get
packets split across multiple `recv/2` calls.