Implement challenge 5
Challenge 5 is challenge 3 (a chat server with a single room) but
reworked as a malicious proxy server to steal your precious bogus coins.

Thanks to the `packet: :line` option this challenge was pretty
straightforward. Both the users and the mob connections only get sent
complete lines; no buffering for either side exists in the
implementation. I spent more time getting the damn regex to work.