Start adding server wiring
This commit adds the groundwork for the server wiring or machinery.

This commit adds much more than I expected, perhaps because I'm not used
to testing this machinery. As a result the code is a bit uglier than
I'd like, e.g. handling named processes. But as there is quite a bit of
wiring going on I hope the testing will pay off.