Handle duplicate identification
This commit handles a minor edge case where a camera sends an
identifying message. Cameras don't need to id themselves so this is an
example of a client going haywire and needs to be treated as such.

The same will need to happen for the dispatcher client when that's in
place.