Process all outstanding messages
This commit fixes an issue with the buffered socket: more than 1 message
may be buffered at a time. For the unidentified client this isn't a
major issue as we only expect 2 types of messages. The camera client
needs to process all types though.

I've also added the socket to the logger for easier debugging.
Surprisingly this needs `inspect(socket)` for the values, which is a bit
clunky but understandable.