Add failing integration test
This commit adds an integration test matching the example listed in the
spec. The first few steps are working but the rest needs much more work.

Running the test is also a pain. This is only a toy project so maybe
that's fine for now. In any case I've designed the test environment to
TCP test functions but the integration test needs to use the real TCP
functions. As a result I'm running the integration test using the dev
environment and getting a ton of warnings.