Register plate detections
This commit adds the first step of handling plate detections. The camera
sees the plate and now the server forwards that on to the core of the
system, the speed checker. We're not yet dealing with any detected
violations.

The integration test notes that we do detect the violation so there's
value in that test already.