Handle identification in Dispatcher
This commit handles my smoke test messages: clients trying to
re-identify, which is not allowed.

I've also added a new way of receiving messages from the buffered socket
to disconnect TCP from messages: the buffered socket can just send all
the messages to a PID, typically the current process. This may mean
other messages being processed first. The only message that may be
problematic is the :tcp_closed message which should be handled first.
It may be simpler to just assume that any send can fail than try to
process messages in a particular order though.