Introduce heart beats
This commit introdues heart beats and the way to register them for a
client (i.e. the "want_heartbeat" message).

This was quite fun to figure out. I thought about linking the socket
more directly to its heartbeat but in the end the heartbeat tells the
client only that the server is operating, not that its particular
client process is operating. So a central heart called `Heart` is
what I ended up with.

The clients are still doing quite a bit for heart beats but each client
deals with it differently so maybe that's fine for now.

There is one known issue I still need to tackle: an interval of 0 should
NOT start a heart beater.