Handle heart beats of 0ms interval
This commit adds support for a special case of the heart beats. An
interval of 0ms should be interpreted as no heartbeat but it should
still register as a heart beater for this client, i.e. the client can't
change their mind and start a heartbeat with a different interval
later.

I made the choice to implement this by starting 1 monitor server for
each 0ms interval heart beater. This is overhead that can be avoided but
it would also complicate things, e.g. a monolithic monitor server that
would need to deregister the heart beater manually. This is only a toy
project so I'll leave it at that.