Use active mode in all clients
This commit changes the unidentified client to use active mode as well.
All clients now use active mode, which is a plus for consistency.

The final message, "plate", is now also handled in the unidentified
client. The other clients need similar work to handle messages they're
not meant to receive.

I could have moved the active mode up to the async server as well to
separate concerns.