Move dispatch to separate server
This commit moves the dispatch functionality (tracking dispatchers and
road combinations, and issuing tickets) to a new server. There are bugs
in this area that I want to fix and it was getting tough to test this
particular area without testing the rest of the road registry, a sure
sign that there are too many responsibilities for this module.

Of course, the dispatch module is very large too. Next up I would break
apart the ticket issuing but I don't think it'll come to that.