Finish challenge 6
Challenge 6 is finally complete. The chasm between 5 and 6 is impressive
but maybe it just shows how inexperienced I am with OTP applications.
Dealing with Phoenix and LiveView feels like a breeze in comparison.

I've experimented with a quite a few modules in this challenge:

* `Registry` to track whether a client has a heartbeat. This would have
  been great for the road registry too.
* ETS to track the many parts of the dispatch state.
* Active mode of :gen_tcp to simplify receiving messages. Unlike
  previous challenges I'm passing the socket around a lot so I learned
  about the "controlling process" the hard way.

All of this was great fun and but if anyone is reading this: beware this
is most likely overengineered. This is not for work so my time is
limited to trying out one or two system designs, and then just going
with to get the challenge completed.